feat: create an install command
